Bytecode: Emit blobs as maps
Because the media-type of a blob is an attribute of the event and not
part of the payload itself, emit a blob as a map with 4 keys-values:

  type: blob
  media-type: MEDIA-TYPE
  total-length: LENGTH
  chunks: CHUNKS ...

Therefore, this commit introduces the new `LTTNG_INTERPRETER_TYPE_BLOB`
to differentiate blobs from arrays and sequences, even though the
interpreter considers them to be identical, the capture serialization
process is done differently.

This commit also adds support for serializing BIN object of the msgpack
format with variable lengths ranging from 1, 2, or 4-byte length blobs.

Because the size of a binary blob is limited to 2^32 - 1 bytes, in
msgpack, the blob is sent with multiple chunks when necessary.

+++ include/lttng/ust-events.h
@@ -217,7 +217,7 @@ struct lttng_ust_type_struct {
 struct lttng_ust_type_fixed_length_blob {
 	struct lttng_ust_type_common parent;
 	uint32_t struct_size;
-	unsigned int length;		/* Num. elems. */
+	unsigned long length;		/* Num. elems. */
 	const char *media_type;		/* Media type, NULL if none. */
 };
